chainwatcher/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
bitarray==3.5.0
certifi==2025.7.9
charset-normalizer==3.4.2
ckzg==2.1.1
cytoolz==1.0.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
exceptiongroup==1.3.0
frozenlist==1.7.0
hexbytes==1.3.1
idna==3.10
iniconfig==2.1.0
multidict==6.6.3
packaging==25.0
parsimonious==0.10.0
pluggy==1.6.0
propcache==0.3.2
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.0.0
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rlp==4.1.0
tomli==2.2.1
toolz==1.0.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
web3==7.12.0
websockets==15.0.1
yarl==1.20.1
